Fix errors

+++ src/main/java/com/github/cargocats/randomjunk/entity/JobAppEntity.java
@@ -1,9 +1,5 @@
 package com.github.cargocats.randomjunk.entity;
 
-import com.github.cargocats.randomjunk.RandomJunk;
-import com.mojang.authlib.minecraft.client.MinecraftClient;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.ai.goal.*;
 import net.minecraft.entity.mob.HostileEntity;
@@ -18,7 +14,7 @@ public class JobAppEntity extends HostileEntity {
     @Override
     protected void initGoals() {
         this.goalSelector.add(0, new SwimGoal(this));
-        this.goalSelector.add(1, new MeleeAttackGoal(this, (double)1.0F, false));
+        this.goalSelector.add(1, new MeleeAttackGoal(this, 1.0f, false));
         this.goalSelector.add(2, new LookAtEntityGoal(this, PlayerEntity.class, 64.0F));
         this.goalSelector.add(2, new LookAroundGoal(this));
